rename pkg to internal (align with other controllers)
