+++ internal/controller/quota/testdata/test-01/onboarding/ns_workspace.yaml
