+++ internal/controller/quota/testdata/test-01/onboarding/qi_workspace_min.yaml
