+++ internal/controller/quota/testdata/test-02/onboarding/qi_normal_delta.yaml
